refactor: Change ID type from uint to uint64 across domain, repository, service, and handler layers

+++ internal/domain/apikey.go
@@ -9,8 +9,8 @@ import (
 
 // APIKey represents authentication credentials for organization API access.
 type APIKey struct {
-	ID             uint       `gorm:"primaryKey"                    json:"id"`
-	OrganizationID uint       `gorm:"not null"                      json:"organization_id"`
+	ID             uint64     `gorm:"primaryKey"                    json:"id"`
+	OrganizationID uint64     `gorm:"not null"                      json:"organization_id"`
 	HashedKey      string     `gorm:"size:255;uniqueIndex;not null" json:"-"` // Hashed, never return raw
 	Label          string     `gorm:"size:100;not null"             json:"label"`
 	CreatedAt      time.Time  `                                     json:"created_at"`
@@ -20,19 +20,19 @@ type APIKey struct {
 // APIKeyRepository defines the interface for API key data operations.
 type APIKeyRepository interface {
 	Create(key *APIKey) error
-	FindByID(id uint) (*APIKey, error)
+	FindByID(id uint64) (*APIKey, error)
 	FindByHashedKey(hashedKey string) (*APIKey, error)
-	ListByOrganizationID(orgID uint) ([]APIKey, error)
-	Revoke(id uint) error
-	Delete(id uint) error
+	ListByOrganizationID(orgID uint64) ([]APIKey, error)
+	Revoke(id uint64) error
+	Delete(id uint64) error
 }
 
 // APIKeyService defines the interface for API key business logic.
 type APIKeyService interface {
-	GenerateKey(orgID uint, label string) (string, error) // Returns the raw key (only shown once)
+	GenerateKey(orgID uint64, label string) (string, error) // Returns the raw key (only shown once)
 	ValidateKey(rawKey string) (*APIKey, error)
-	GetByID(id uint) (*APIKey, error)
-	ListByOrganizationID(orgID uint) ([]APIKey, error)
-	RevokeKey(id uint) error
-	DeleteKey(id uint) error
+	GetByID(id uint64) (*APIKey, error)
+	ListByOrganizationID(orgID uint64) ([]APIKey, error)
+	RevokeKey(id uint64) error
+	DeleteKey(id uint64) error
 }
